requirements.txt
opencv-python
numpy
mediapipe
pandas
pyyaml
sanic
matplotlib
pandas
jupyter
tqdm
pygame
seaborn